refactor: create GST return log instead of GSTR-1 log
