+++ india_compliance/gst_india/doctype/gst_return_log/test_gst_return_log.py
@@ -5,5 +5,5 @@
 from frappe.tests.utils import FrappeTestCase
 
 
-class TestGSTR1FiledLog(FrappeTestCase):
+class TestGSTReturnLog(FrappeTestCase):
     pass
